add solver: succinct small progress measures
This adds an implementation the "succint small progress measures"
algorithm proposed by Jurdzinski/Lazic in https://arxiv.org/abs/1702.05051.

The special comparisons for bitstrings and derived bounded adaptive
counters are done explicitly in ocaml, which is slow.

So far, this only derives a winning strategy for player 1
and sets the strategy for player 0 to a trivial bogus strategy.